3sxtra

+++ src/sf33rd/Source/Game/demo/demo01.h
@@ -1,9 +1,16 @@
+/**
+ * @file demo01.h
+ * @brief Public API for title screen sequences.
+ *
+ * Part of the demo module.
+ */
+
 #ifndef DEMO01_H
 #define DEMO01_H
 
 #include "types.h"
 
-s32 Title_At_a_Dash();
+s16 Title_At_a_Dash();
 s16 Title();
 
 #endif
